Day 3 Training
finished Heli set up
fixed only see Instructors role on training
set up for costume vic loadouts
