+++ maps/vn_khe_sanh/map_config/teams_append.hpp
@@ -1,0 +1,1 @@
+// No map-specific teams for this map.
